+++ Heros/Block.h
@@ -6,7 +6,7 @@ class Block : public GameObject {
 public:
     Block(Texture);
     Block(Texture, Vector2);
-    void Update(float, Keys);
+    void Update(double, Keys);
     void Render(Viewport &);
 private:
     Texture texture_;
